fix(linux): Rename dir for consistency
Use shorter OS subpackage names for a consistent experience.
