+++ vgamepad/__init__.py
@@ -4,4 +4,4 @@ from vgamepad.win.vigem_commons import VIGEM_TARGET_TYPE, XUSB_BUTTON, DS4_BUTTO
 if os.name == 'nt':
     from vgamepad.win.virtual_gamepad import VX360Gamepad, VDS4Gamepad
 else:
-    from vgamepad.linux.virtual_gamepad import VX360Gamepad, VDS4Gamepad
+    from vgamepad.lin.virtual_gamepad import VX360Gamepad, VDS4Gamepad
